byte code instructions can fit in a byte
This isn't particularly important.  It just struck me that if we're
going to be byte-thrifty, might as well go all the way.

+++ src/sksl/SkSLByteCode.h
@@ -19,7 +19,7 @@ namespace SkSL {
 class  ExternalValue;
 struct FunctionDeclaration;
 
-enum class ByteCodeInstruction : uint16_t {
+enum class ByteCodeInstruction : uint8_t {
     // B = bool, F = float, I = int, S = signed, U = unsigned
 
     kAddF,  // N
